requirements.txt
gradio
python-dotenv
flask
langchain
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
onnxruntime
pyttsx3
